+++ tests/test_stats.py
@@ -3,14 +3,14 @@ from scipy.ndimage import gaussian_filter
 import os
 import inspect
 import sys
-from nose.tools import nottest
+import pytest
 
 LOCATION = "/".join(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))).split("/")[:-1])
 sys.path.insert(0, LOCATION)
 
 from powerbox import PowerBox
 
-@nottest                  # this is not passing to desired tolerance at this point... not sure if this is a problem. It's not systematic.
+@pytest.mark.skip(reason="this is not passing to desired tolerance at this point... not sure if this is a problem. It's not systematic.")
 def test_resolution():
     var = [0]*6
     for i in range(6):